WindowImpl, MultiTypedef, window timers, more WebIDL adapting logic,

+++ web-k-core/src/main/java/com/earnix/webk/script/html/WindowOrWorkerGlobalScope.java
@@ -40,11 +40,11 @@ public interface WindowOrWorkerGlobalScope {
     String atob(@DOMString String data);
 
     // timers
-    long setTimeout(TimerHandler handler, @Optional @DefaultLong(0) int timeout, Object... arguments);
+    int setTimeout(TimerHandler handler, @Optional @DefaultLong(0) int timeout, Object... arguments);
 
     void clearTimeout(@Optional @DefaultLong(0) int handle);
 
-    long setInterval(TimerHandler handler, @Optional @DefaultLong(0) int timeout, Object... arguments);
+    int setInterval(TimerHandler handler, @Optional @DefaultLong(0) int timeout, Object... arguments);
 
     void clearInterval(@Optional @DefaultLong(0) int handle);
